added utility to nanolistener_cons_out_filt.py
nanolistener_cons_out_filt.py is an utility script n python to filter chunks and to merge NanoListener consumers output with filtering options on mod.base to select (if requested), the kmer taget max and min length, the output chunks length and padding options, the balancing strategy, n of chunks and the train-test splist ratio. It produce a training and a testing dataset to be used to train NanoSpeech models.